% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvault-package.R
\docType{package}
\name{pvault-package}
\alias{pvault}
\alias{pvault-package}
\title{pvault: Generate Encripted Project Vaults}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Create encrypted project vaults and easily store arbirtary R
  objects to it.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GregorDeCillia/pvault}
  \item Report bugs at \url{https://github.com/GregorDeCillia/pvault/issues}
}

}
\author{
\strong{Maintainer}: Gregor de Cillia \email{de.cillia.gregor@gmail.com}

}
\keyword{internal}
