% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock.R, R/vault.R
\name{lock_vault}
\alias{lock_vault}
\alias{unlock_vault}
\alias{vault}
\alias{create_vault}
\alias{delete_vault}
\title{Manage vaults}
\usage{
lock_vault(password)

unlock_vault(password)

create_vault()

delete_vault()
}
\arguments{
\item{password}{A password to lock/unlock the vault}
}
\description{
Create and delete a vault for the current project
}
